+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Let the Google sheet importer sort media files related to each organization
   or course in their specific folder in Django filer,
+- Make the superuser field readonly for non superusers.
 
 ### Fixed